Add files via upload
A bank is a financial institution that accepts deposits from the public and creates a demand deposit while simultaneously making loans. A deposit is a financial term that means money held at a bank and it can refer to a portion of the money used as security or collateral for the delivery of a good. So, how many clients are deposited in a bank daily matters a lot? In a bank, every client has their previous transition record and their bank records. By using this data, we can use machine learning to predict a client will deposit in a bank or not. It will help a lot in a bank to increase their profit. In this paper, we examine the best machine learning technique to classify. To get the best result we used a variety of metrics, including classification accuracy and required dataset so that we can get real-time predict result correctly. So, I used the Decision Tree technique to predict a client is deposited in a bank or not.